server/requirements.txt
fastapi==0.120.3
uvicorn[standard]==0.38.0
sqlmodel==0.0.27
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pydantic-settings==2.11.0
alembic==1.13.1
PyJWT==2.10.1
email-validator==2.1.0
python-multipart==0.0.9
passlib==1.7.4
pydantic==2.12.4
pydantic_core==2.41.5
argon2-cffi==25.1.0
